requirements.txt
langchain
langchain-community
crewai
crewai_tools
python-dotenv
ipykernel